Create gdi_des.md
The project prepares Gender Development Index based on Socio-Economic Indicators as given in NFHS 4 Data for all States. It clusters the states, visualizes the pattern on heatmap and conducts PCA to determine the factors that describe the patterns in data. Based on these principal components we deduce the major factors driving GDI