Add files via upload
The research paper was written collaborating with People's Initiative a not-for-profit focusing on social sector research. The paper was published in International Journal for Social Sciences. The excel file is the raw analysis, testing, clustering. I have subsequently coded in Python the same model. But it was originally in Excel. 
The journal ISSN:2249-6637
The research attempted to establish a relation between IMR and key social indicators like female literacy, institutional birth percentages. The model was a linear model.